Add `y` to second pacman update call.
If the first "core" update updated pacman with a new conf that had a new
sync db, that db would not be present and the second call would error
out.  Add s`y`nc to second update call so any missing DBs will be
downloaded, at the expense of pinging the servers to see if any of the
other databases were updated in the meantime.