Add variable to put runner into runner group.
Runner groups are relatively recently available to free users, and can
be restricted as to what repos and workflows (including branches) can
use them.

Also, remove the logic for an ephemeral runner that shuts down after one
job: this was an aspirational thing when I started this repo, but is not
what I'm doing in practice.  By the time I could architect something to
make it possible, GitHub would probably have made arm64 runners
available to free users anyway, rendering the effort moot.